fix: 수익 반환 자료형 변경

+++ src/main/java/blackjack/gambler/Gambler.java
@@ -55,7 +55,7 @@ public abstract class Gambler {
         return cards;
     }
 
-    public double getBetAmount() {
-        return betAmount;
+    public int getBetAmount() {
+        return (int) betAmount;
     }
 }
